+++ src/view/video/smartGenerate.vue
@@ -266,8 +266,12 @@ export default {
       sessionStorage.setItem("copy_list", JSON.stringify(this.copy_list));
     },
     removeCopy(index) {
-      this.copy_list.splice(index, 1);
-      sessionStorage.setItem("copy_list", JSON.stringify(this.copy_list))
+      this.$confirm('确认删除该文案吗？', '提示').then(() => {
+        this.copy_list.splice(index, 1);
+        sessionStorage.setItem("copy_list", JSON.stringify(this.copy_list))
+      }).catch((err) => {
+        this.$message({type: 'info', message: '已取消删除'});
+      });
     },
     nextStep() {
       if (this.copy_list.length === 0) {
